Revamped AI System
This could be considered V2 of the previous program, but even this is incomplete as of now. I changed the NN to an actor/critic system, I added a lot more functionality to the screen capturing, added a functional reward system, and pieced everything together to finalize and get a basic RL AI working.